this is the patch from trunk: lint

+++ net.c
@@ -137,9 +137,9 @@ ldns_send(ldns_pkt **result, ldns_resolver *r, ldns_pkt *query_pkt)
 		gettimeofday(&tv_e, NULL);
 
 		if (reply) {
-			ldns_pkt_set_querytime(reply,
-				(uint32_t)((tv_e.tv_sec - tv_s.tv_sec) * 1000) +
-				(uint32_t)(tv_e.tv_usec - tv_s.tv_usec) / 1000);
+			ldns_pkt_set_querytime(reply, (uint32_t)
+				((tv_e.tv_sec - tv_s.tv_sec) * 1000) +
+				(tv_e.tv_usec - tv_s.tv_usec) / 1000);
 			ldns_pkt_set_answerfrom(reply, ns_rand_array[i]);
 			ldns_pkt_set_when(reply, ctime((time_t*)&tv_s.tv_sec));
 			ldns_pkt_set_size(reply, reply_size);
